move validation code
